Rename/Deprecate RBL for CIMLK
Put this naming change to avoid having future updates to IMAER-Java in case only change would require updating library only to have this rename.
